+++ src/main/java/com/sofa/linkiving/domain/link/repository/LinkRepository.java
@@ -69,5 +69,5 @@ public interface LinkRepository extends JpaRepository<Link, Long> {
 		@Param("member") Member member
 	);
 
-	int countByMemberAndIsDeleteFalse(Member member);
+	Long countByMemberAndIsDeleteFalse(Member member);
 }
